// generated/src/app/api-models/cap-util-pool-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapUtilPoolWork {
  seqCcalcId: number;
  seqCapUtilPoolWork: number;
  seqCapPoolId: number;
  applyTo: string;
  capStoplossId: string;
  seqCapElgcntWork: number;
  seqMembId: number;
  applicableAmt: number;
  applicableAmtUpToDate: Date;
  stoplossAdjDate: Date;
  stoplossThreshold: number;
  companyCode: string;
  glRefCode: string;
  postingRule: string;
}